+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Added
+- Add Swift Concurrency support
+
+### Removed
+- Drop support for iOS 11 and iOS 12
+- Drop support for Swift below 5.7
+
 ## [11.5.0] - 2023-09-19
 
 ### Updated